requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langgraph
langgraph-checkpoint-sqlite
chromadb
python-dotenv
pydantic
pandas
tqdm
grandalf
streamlit